+++ internal/server/transport/tcpmux.go
@@ -340,7 +340,7 @@ func (s *TcpMuxTransport) channelHandshake(g *tcpMuxGen) {
 }
 
 func (s *TcpMuxTransport) channelHandler(g *tcpMuxGen) {
-	ticker := time.NewTicker(livenessBeat(s.config.Heartbeat))
+	ticker := newLivenessTicker(s.config.Heartbeat)
 	defer ticker.Stop()
 
 	// Channel to receive the message or error